test-base.cfg: remove lovely.buildouthttp extension.
The lovely.buildouthttp extension depends on six's urllib port,
which is no longer available in the current six version.

We also do not really use the lovely.buildouthttp extension,
since it does not work properly with never buildout versions.

The authentication.cfg should be used instead, which is based
on isotoma.buildout.basicauth.