Fix implicity dependency problem (CMFPlacefulWorkflow).
plone.app.testing now imports from Products.CMFPlacefulWorkflow, which
is often not installed because no explicity dependency is declared.
We now always install the Plone egg in order to have the full Plone stack ready.

The problem has been already fixed for the 4.3.x test setup, but 4.1.x and 4.2.x needs the fix as well.