Remove plone.recipe.precompiler from production.cfg:
If the buildout directory is the same as a development egg
(which is the case for policy packages checked out from source
that include the buildout.cfg alongside the code), it will scan
pretty much the entire deployment directory during buildout.

For deployments with many files, such as a reasonably large
blobstorage, this leads to a severe performance hit while
running buildout.