add autostart options for HttpOk and Memmon.
Sometimes we want to not use HttpOk or Memmon.
By setting the autostart option of those eventlisteners we can tell
whether the listeners should automatically be started when starting the
supervisor daemon, e.g. when the machine is rebootet.

Before, we didn't set autostart and supervisor used ``true`` as default.
Now, we explicitly set it to ``true`` and have the option to change it.

Example:

    [buildout]
    supervisor-memmon-autostart = false
    supervisor-httpok-autostart = false