extract supervisor eventlisteners into variables.
In #48 we had to revert the httpok / memmon autostart feature because we
might have older plone.recipe.supervisor versions which do not support
this syntax.

On order to be able to not have HttpOk (or Memmon) eventlisteners we
move the listener definition into separate buildout variables, one for
each listener, and then just join the variables.

This gives us the possibility to overwrite the "eventlisteners" option
and join only the variables of the wanted event listeners, skipping
those which we do not want.

This solves the initial use case: to not have HttpOk eventlisteners run
or configured.