supervisor-haproxy event listener configuration.
The `supervisor-haproxy` event listener tells haproxy to remove
backends / add backends to the load balancing when supervisor detects
instances to be stopping and starting.

The `production.cfg` contains the default configuration variables such
as the haproxy-socket or the program mapping variable but it does not
set up the event listener. The zeoclients-config-files then extend the
program configuration dynamically.

For actually enabling the feature the `haproxy.cfg` must be extended.
This way we avoid accidental activation.