Fix self-update by not setting "newest" option.
The "newest=false" option has the effect that buildout does no longer
self-update when another version is pinned than istalled.

This may be quite a problem especially when we rely on having at least a
certain buildout version.

Since we use allow-picked-versions=false we can remove the
"newest=false" option since this means all versions have to be pinned
anyways.