plone-development: install XML formatting scripts.
This adds a `bin/format-xml` and a `bin/format-all-xmls`.

The `bin/format-xml` can format one or many XML files.
It reindents all the XML nodes, but it also adds newlines between
attributes so that XML nodes with many attributes, such as `<configure>`
tags, look nicely.

The `bin/format-all-xmls` simply looks for all `.zcml` and `.xml` files
within the projects and calls `bin/format-xml` for each.
It ignores everything within pull-requests as we have no value in
reformatting those.