Fix _strptime race condition in p.r.zope2instance based instances:
There is an unfixed race condition in Python around the very first
import of the _strptime C module: That import isn't thread safe,
and if two threads happen to trigger this lazy import at the same
time, the interpreter crashes.

See http://bugs.python.org/issue7980

The workaround for this issue is to import _strptime before any
threads are started, while the application is still single-threaded.
We therefore address this here by triggering the import in the
initialization section of p.r.zope2instance based instances.