redis: revert mode change causing side effects.
Reverts part of 0b33899

The problem is that by changing mode in ththe ``redis-download``
section, buildout re-ran the part and downloaded and extracted a new
redis archive.
But since this happened in an "update", the redis-build section
did not rebuild because of ``on_update = false``.
This caused redis to not be built and the ``redis-server`` executable
was missing.