Add testing variables for package extras.
The new variable ``buildout:package-extras`` can be used in the testing
buildout for installing and testing an optional feature of the package
under test.

This change also introduces the variable ``buildout:test-extras``,
which combines the ``buildout:package-extras`` and the default "tests"
extras. It makes it easier to change the behavior for special use cases.