Fix redis make error on fresh download.
The redis-download-part downloads a fresh source copy of redis, where
redis-build-part compiles the source.

redis-build's on_update was set to false in order to speed up updates
where redis is already installed and does not change.

But on certain changes (such as a changed redis buildout file) buildout
will reinstall the redis-download part while only running "update" on
the redis-build part.
Since redis-build will skip we will end up with an uncompiled redis
source.

In order to solve this problem we change on_update to true, so that the
redis-build part is always executed.
For still skipping on an already installed redis, we just probe for the
src/redis-server file and only run "make" when the file is missing.