Let only instance0 eager-load chameleon templates by default.
We've realized that it is probably a bad idea to let all instances
eager-load (=precompile) chameleon templates.
It can get messy and lead to errors when multiple instances are
precompiling at the same time.

In order to only precompile on instance0 by default, the
`chameleon-eager` configuration option for production is split into a
`chameleon-eager-default` option (used for instances 1-4) and a
`chameleon-eager-instance0` option (used for instance0 only).
For publisher instances we disable eager loading because they never
render things.

Since all instance parts copy from `instance0`, the option must be
configured in each instance part with another option.

This change has the side effect that we now always configure the
`CHAMELEON_EAGER` enviroment variable, even when `chameleon.cfg` is
not extendend.