Use distinct chameleon cache directories per instance.
We've had chameleon compilation errors which may have been caused by
using a shared chameleon cache.
(see https://community.plone.org/t/chameleon-typeerror-dict-object-is-not-callable/5391)

Therefore we're changing buildout to configure a separate chameleon
cache directory for each instance.

The directories must exist and are not created by Chameleon.
For easier understanding and handling we just create directories for
all instances we now in ftw-buildouts, no metter whether they are
actually used or not.