Ensure chameleon cache directory variable for all tests.
The chameleon.cfg is usually not extended by test buildout configs,
thus it does not ensure the CHAMELEON_CACHE environment variable on
the CI server.

The CI server used to set the directory to one globally shared
directory, which is a bad idea because of concurrency issues.

Therefore we now always provide a chameleon cache directory, if not
yet set, so that each test has an own chameleon cache directory and we
have no concurrency issues.