Update allow-hosts to make redirects to  pypi.org work:
pypi.python.org started redirecting a fraction of its traffic
to pypi.org in preparation for the final rollout of the new
codebase.

This means, in order for downloads to be allowed from the new
location, both pypi.org (the package index) and
files.pythonhosted.org (where the actual packages are hosted)
need to be whitelisted in allow-hosts.