Prefer final releases in testing.
Pre-releases of any dependencies may be submitted to pypi.
Those releases may be broken and not installable.
If we are not pinning anything, the "prefer-final=false" setting will install those broken releases,
making our tests fail.
We should not test against non-final third-party releases.
If we really want to test against a non-final version we can still pin in in the versions.cfg.