Add separate test-versions files for the Plone major versions.
Since we have a lot of packages which test against a specific Plone
version but also against the newest third party dependencies, we run
into problems when the third part dependencies stop supporting our
Plone version in a new release; we need to pin it down.

In order to do this in a controlled manner, this change introduces
split version files for Plone 4 and 5 which all Plone test base
configs extend from.

This will allow us to centrally propose different version pinnings for
a package in Plone 4 and 5, while leaving the possibility to the
package to override this pinning.