Constrain PAM to a Plone 5.1 compatible version in test-versions-plone-5.cfg:
p.a.multilingual >= 5.4 is not compatible with Plone 5.1 any more, and in
version 5.4.0 actually introduces a hard dependency on Plone 5.2.

Because Plone 5.1.x currently is the only Plone 5 version we test against,
we constrain PAM to a version that works with it here. At a later point,
the test-versions-plone-5.cfg file should probably be split up into
5.1 and 5.2 variants.