Provide proposed versions fix for testing.
The production.cfg now provides proposed versions and policy projects
should set the zc.buildout and setuptools versions with
proposed-versions according to the docs.

But when setting the versions in the versions.cfg by using the
proposed-versions variables, these variables are missing in testing
because they are defined in production.cfg, which is not used there.

Therefore we need to provide proposed versions in the
test-versions.cfg too.

We cannot use the variables in the same file in the versions section
because buildout will not substitute the variables when the
mr.developer extension is used in certain ways.
Thus we must define the versions for zc.buildout and setuptools twice
in test-versions.cfg.