Factor out Solr ZCML into its own buildout variable:
This is necessary to be able to easily override (completely
replace) the Solr ZCML in a buildout extending from solr-v2.cfg.

More specifically, we need to add the "upload_blobs" ZCML attribute
for deployments with Solr master/slave replication. But because this
ZCML attribute doesn't exist yet in older versions of ftw.solr, it
can't simply be added to ftw-buildouts with a default of "false".