Pin package "inflection" to <0.4.0 for python2.
The package has dropped python2 support in version 0.4.0 according to
https://github.com/jpvanhal/inflection/releases/tag/0.4.0. We should
use a lower version when running against python2.

Interstingly only version 0.5.0 started to really break things for me.
So if someone needs 0.4.0 for some reason we might be able to change
the version constraint. But i'd prefer to go with what the package
announced officially, so i'd suggest to try this change first and see
it if breaks something for us.