jenkins-run.py: Fix bootstrapping of buildouts with Python 3:
zc.buildout >= 3 never was compatible with Python 2.7, and we therefore had
to explicitly use the last 2.x version when bootstrapping.

We only did this conditionally for Python 2.7 though. Now it seems that
bootstrapping with zc.buildout >= 3 has also stopped working for Python 3,
because of an error like this during bootstrap:

```
pkg_resources.DistributionNotFound: The 'wheel' distribution was not found and
is required by zc.buildout
```

We therefore unconditionally use buildout 2.13.8 for bootstrapping on Jenkins.